requirements.txt
tk
opencv-python
Pillow
numpy
joblib
pandas
bcrypt
ultralytics
deepface